refactor for new internal interface
